Move strip attributes logic to strip function.
